Max Depth, 64-bit Keys, Improve Mate Detection

+++ src/misc.h
@@ -46,7 +46,7 @@ inline TimePoint now() {
 
 template<class Entry, int Size>
 struct HashTable {
-  Entry* operator[](Key key) { return &table[(uint32_t)key & (Size - 1)]; }
+  Entry* operator[](Key key) { return &table[key & (Size - 1)]; }
 
 private:
   std::vector<Entry> table = std::vector<Entry>(Size);
